the-powder-toy.py: Initial version
Add an upstream watcher for the game "The Powder Toy".
http://powdertoy.co.uk/

The download link does not contain the version. The version
is only text on the page. This upstream watcher gets the version
and creates a download link for it. This link does not exist so
you get a 404. But at least you know where to look for the real
download link because the version is prefixed with the homepage URL.