the-powder-toy.py: Scan GitHub for upstream versions
From now on each new release is tagged on GitHub.
Unfortunately, the tag name is not based on the version but
on the build number. So we still need this upstream watcher
to pick the version from the tag description instead.
This also returns an invalid URL but just removing the
version brings you to the tag download page on GitHub for
the PowderToypp project.