tide.py: New upstream watcher for tIDE (a java IDE)
The site http://tide.olympe.in/main.htm announces
the new release with a version but this version is not
part of the download link. The link just points to
a tide.jar. So this watcher gets the version and appends
it to "http://tide.olympe.in/". Again this gives a 404
but you will have the place to look for the download link.