sandboxgamemaker.py: Initial version
The download links of Platinum Arts Sandbox Free Game Maker
just contain an incrementing number which has nothing to do
with the actual version. This upstream watchers tries to find
the version in the description of the download on the site.
The generated link will 404 of course but will also give you
the URL to the download site.