the-powder-toy.py: Get version from homepage an point to GitHub
Unfortunately, the project does not use version numbers but
build numbers as their git tags. So we have to search the version
on the homepage and point to GitHub. This will 404 but at least
we are on the correct site for the source code.
The homepage seems to make some crazy OS guessing:
In my browser there is: "Version 85.0 for Linux" but with
python there is just "Version 85.0 for " (probably because
of the missing user agent).