calibre.py: Watcher for calibre because sf is not maintained
According to the developer the source should not be downloaded
from Sourceforge but from their own servers. (LP: #1286535)
Unfortunately, the directory structure makes it hard to find the
correct version. But they provide a fixed URL which redirects
to the latest source tarball. So get the location of this tarball
and print the complete URL which contains version information.